add camera class into components module, fix somebugs

+++ Pod/Component.cpp
@@ -1,6 +1,6 @@
 #include "Component.h"
 
-Component::Component(int _x, int _y)
+Component::Component(float _x, float _y)
 {
 	this->position = new D3DXVECTOR3(_x, _y, 0);
 }
@@ -20,12 +20,12 @@ D3DXVECTOR3* Component::getPosition()
 	return this->position;
 }
 
-int Component::getX()
+float Component::getX()
 {
 	return this->position->x;
 }
 
-int Component::getY()
+float Component::getY()
 {
 	return this->position->y;
 }
@@ -35,22 +35,22 @@ void Component::setPosition(D3DXVECTOR3* _position)
 	this->position = _position;
 }
 
-void Component::setX(int _x)
+void Component::setX(float _x)
 {
 	this->position->x = _x;
 }
 
-void Component::setY(int _y)
+void Component::setY(float _y)
 {
 	this->position->y = _y;
 }
 
-void Component::plusX(int _x)
+void Component::plusX(float _x)
 {
 	this->position->x += _x;
 }
 
-void Component::plusY(int _y)
+void Component::plusY(float _y)
 {
 	this->position->y += _y;
 }
